+++ pkg/analyzer/lib/src/error/codes.g.dart
@@ -257,11 +257,31 @@ class CompileTimeErrorCode extends AnalyzerErrorCode {
         "Try changing the augmentation to match the declaration kind.",
   );
 
+  static const CompileTimeErrorCode AUGMENTATION_TYPE_PARAMETER_BOUND =
+      CompileTimeErrorCode(
+    'AUGMENTATION_TYPE_PARAMETER_BOUND',
+    "The augmentation type parameter must have the same bound as the "
+        "corresponding type parameter of the declaration.",
+    correctionMessage:
+        "Try changing the augmentation to match the declaration type "
+        "parameters.",
+  );
+
   static const CompileTimeErrorCode AUGMENTATION_TYPE_PARAMETER_COUNT =
       CompileTimeErrorCode(
     'AUGMENTATION_TYPE_PARAMETER_COUNT',
     "The augmentation must have the same number of type parameters as the "
         "declaration.",
+    correctionMessage:
+        "Try changing the augmentation to match the declaration type "
+        "parameters.",
+  );
+
+  static const CompileTimeErrorCode AUGMENTATION_TYPE_PARAMETER_NAME =
+      CompileTimeErrorCode(
+    'AUGMENTATION_TYPE_PARAMETER_NAME',
+    "The augmentation type parameter must have the same name as the "
+        "corresponding type parameter of the declaration.",
     correctionMessage:
         "Try changing the augmentation to match the declaration type "
         "parameters.",